DailyChallenge5Apr
﻿

The program must accept an integer matrix of size NxN as the input. The program must print the product of the sum of integers in the diagonals (left and right diagonals) of the matrix as the output.
Boundary Condition(s):
3 <= N <= 50
Input Format:
The first line contains N.
The next N lines, each containing N integers separated by a space.
Output Format:
The first line contains the product of the sum of integers in the diagonals (left and right diagonals) of the matrix.
Example Input/Output 1:
Input:
4
6792
2441
9391
1399
Output: 280
Explanation:
The sum of integers in the left diagonal is 28 (6 + 4 + 9 + 9).
The sum of integers in the right diagonal is 10 (2 + 4 +3 +1). So the product of 28 and 10 is 280.
Hence the output is 280
Example Input/Output 2:
Input:
3
123
456
789
Output:
225